add check file exist

+++ products.py
@@ -1,25 +1,33 @@
+import os
+
 products = []
-with open('products.csv', 'r', encoding='utf-8') as f:
-	for line in f:
-		if '商品名稱,價格' in line:
-			continue
-		name, price = line.strip().split(',')
+if os.path.isfile('products.csv'):
+	#讀取檔案
+	with open('products.csv', 'r', encoding='utf-8') as f:
+		for line in f:
+			if '商品名稱,價格' in line:
+				continue
+			name, price = line.strip().split(',')
+			products.append([name,price])
+	print(products)
+
+else:
+	#讓使用者輸入
+	while True:
+		name = input('請輸入商品名稱： ')
+		if name == 'q':
+			break
+		price = int(input('請輸入商品價格： '))
 		products.append([name,price])
-print(products)
+	print(products)
 
-while True:
-	name = input('請輸入商品名稱： ')
-	if name == 'q':
-		break
-	price = int(input('請輸入商品價格： '))
-	products.append([name,price])
-print(products)
-
-for p in products:
-	print(p[0], '的價格是', p[1])
-
-
-with open('products.csv', 'w', encoding='utf-8') as f:
-	f.write('商品名稱' + ',' + '價格' + '\n')
+	#印出所有購買清單
 	for p in products:
-		f.write(p[0] + ',' + str(p[1]) + '\n')
+		print(p[0], '的價格是', p[1])
+
+
+	#寫入檔案
+	with open('products.csv', 'w', encoding='utf-8') as f:
+		f.write('商品名稱' + ',' + '價格' + '\n')
+		for p in products:
+			f.write(p[0] + ',' + str(p[1]) + '\n')
